Feedback addressed
